+++ styleguide/component.stories.js
@@ -17,6 +17,7 @@ import wrapper from './component-wrapper';
 
 import Buttons from './components/buttons';
 import { Pagination, PagingInfo } from './components/pagination';
+import RelatedContent from './components/related-content';
 
 // Haml setup
 // Import all the templates...
@@ -81,6 +82,7 @@ export default {
 export const buttons = () => Buttons();
 export const pagination = () => Pagination();
 export const pagingInfo = () => PagingInfo();
+export const relatedContent = () => RelatedContent();
 
 // The haml components. Thet will be sorted alphabetically so the order here is not important.
 export const search = () => renderHamlTemplate('Search', tSearch, 'search');